Remove GenreWordConverterFactory
