+++ src/content-script-action/for-detail-search-page.ts
@@ -1,6 +1,6 @@
 
 import { ContentScriptAction } from "../content-script-action";
-import { GenreWordConverterFactory } from "../core/genre-word-converter-factory";
+import { GenreWordConverter } from "../core/genre-word-converter";
 import { GenreWordReplacerFactory } from "../core/genre-word-replacer-factory";
 import { GenreWordReplaceTargetPage } from "../core/genre-word-replace-target-page";
 import { Message } from "../message";
@@ -91,10 +91,9 @@ function doReplaceGenreWords() {
         const conversionMap  = (results[0].data as MessageDataGetConversionMapResponse ).conversionMap;
         const conversionMode = (results[1].data as MessageDataGetConversionModeResponse).conversionMode;
     
-        const wordConverterFactory = new GenreWordConverterFactory();
-        const wordReplacerFactory  = new GenreWordReplacerFactory();
-        const wordConverter        = wordConverterFactory.createGenreWordConverter(conversionMap, conversionMode);
-        const wordReplacer         = wordReplacerFactory.createGenreWordReplacer(GenreWordReplaceTargetPage.DetailSearchPage);
+        const wordConverter       = new GenreWordConverter(conversionMap, conversionMode);
+        const wordReplacerFactory = new GenreWordReplacerFactory();
+        const wordReplacer        = wordReplacerFactory.createGenreWordReplacer(GenreWordReplaceTargetPage.DetailSearchPage);
     
         wordReplacer.replaceGenreWords(document, wordConverter);
     })
